# Speedup notchfilter