Previous timings:
  time_rereference : 1.57ms
  time_resample : 2.2ms
  time_notchfilter : 15.91ms
  time_feature_norm : 0.18ms
  time_projection : 4.07ms
  time_fft : 0.3ms

Now with precomputed notch filter:
  time_rereference : 1.79ms
  time_resample : 2.5ms
  time_notchfilter : 3.7ms
  time_feature_norm : 0.28ms
  time_projection : 4.99ms
  time_fft : 0.44ms